--- backend/src/main/java/com/woowacourse/momo/CustomNamingStrategy.java
@@ -1,16 +1,0 @@
-package com.woowacourse.momo;
-
-import org.hibernate.boot.model.naming.Identifier;
-import org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl;
-import org.hibernate.engine.jdbc.env.spi.JdbcEnvironment;
-
-public class CustomNamingStrategy extends PhysicalNamingStrategyStandardImpl {
-
-    private static final String PREFIX = "momo_";
-
-    @Override
-    public Identifier toPhysicalTableName(Identifier name, JdbcEnvironment context) {
-        Identifier prefixedTableName = new Identifier(PREFIX + name.getText().toLowerCase(), name.isQuoted());
-        return super.toPhysicalTableName(prefixedTableName, context);
-    }
-}
