+++ backend/src/main/java/com/woowacourse/momo/category/controller/CategoryController.java
@@ -1,0 +1,26 @@
+package com.woowacourse.momo.category.controller;
+
+import java.util.List;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import lombok.RequiredArgsConstructor;
+
+import com.woowacourse.momo.category.service.CategoryService;
+import com.woowacourse.momo.category.service.dto.response.CategoryResponse;
+
+@RequiredArgsConstructor
+@RequestMapping("/api/categories")
+@RestController
+public class CategoryController {
+
+    private final CategoryService categoryService;
+
+    @GetMapping
+    public ResponseEntity<List<CategoryResponse>> findAll() {
+        return ResponseEntity.ok(categoryService.findAll());
+    }
+}
