+++ backend/src/main/java/com/woowacourse/momo/category/service/CategoryService.java
@@ -1,0 +1,29 @@
+package com.woowacourse.momo.category.service;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import lombok.RequiredArgsConstructor;
+
+import com.woowacourse.momo.category.domain.Category;
+import com.woowacourse.momo.category.domain.CategoryRepository;
+import com.woowacourse.momo.category.service.dto.response.CategoryResponse;
+
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+@Service
+public class CategoryService {
+
+    private final CategoryRepository categoryRepository;
+
+    public List<CategoryResponse> findAll() {
+        List<Category> categories = categoryRepository.findAll();
+
+        return categories.stream()
+                .map(CategoryResponse::from)
+                .collect(Collectors.toList());
+    }
+}
