--- backend/src/main/java/com/woowacourse/momo/group/controller/CategoryController.java
@@ -1,26 +1,0 @@
-package com.woowacourse.momo.group.controller;
-
-import java.util.List;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import lombok.RequiredArgsConstructor;
-
-import com.woowacourse.momo.group.service.CategoryService;
-import com.woowacourse.momo.group.service.dto.CategoryResponse;
-
-@RequiredArgsConstructor
-@RequestMapping("/api/categories")
-@RestController
-public class CategoryController {
-
-    private final CategoryService categoryService;
-
-    @GetMapping
-    public ResponseEntity<List<CategoryResponse>> findAll() {
-        return ResponseEntity.ok(categoryService.findAll());
-    }
-}
