--- backend/src/main/java/com/woowacourse/momo/group/domain/Schedules.java
@@ -1,45 +1,0 @@
-package com.woowacourse.momo.group.domain;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import javax.persistence.CascadeType;
-import javax.persistence.Embeddable;
-import javax.persistence.OneToMany;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Embeddable
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Schedules {
-
-    @OneToMany(mappedBy = "group",
-            cascade = {CascadeType.PERSIST, CascadeType.REMOVE},
-            orphanRemoval = true)
-    private List<Schedule> value = new ArrayList<>();
-
-    public Schedules(List<Schedule> value) {
-        this.value = value;
-    }
-
-    public Schedules(List<Schedule> value, Group group) {
-        this(value);
-        belongTo(group);
-    }
-
-    private void belongTo(Group group) {
-        value.forEach(v -> v.belongTo(group));
-    }
-
-    public void clear() {
-        value.clear();
-    }
-
-    public void add(Group group, Schedule schedule) {
-        this.value.add(schedule);
-        schedule.belongTo(group);
-    }
-}
