+++ backend/src/main/java/com/woowacourse/momo/group/domain/group/GroupRepository.java
@@ -1,0 +1,6 @@
+package com.woowacourse.momo.group.domain.group;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface GroupRepository extends JpaRepository<Group, Long> {
+}
