--- backend/src/main/java/com/woowacourse/momo/group/repository/GroupRepository.java
@@ -1,8 +1,0 @@
-package com.woowacourse.momo.group.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.woowacourse.momo.group.domain.Group;
-
-public interface GroupRepository extends JpaRepository<Group, Long> {
-}
