--- backend/src/main/java/com/woowacourse/momo/group/service/CategoryService.java
@@ -1,29 +1,0 @@
-package com.woowacourse.momo.group.service;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import lombok.RequiredArgsConstructor;
-
-import com.woowacourse.momo.group.domain.Category;
-import com.woowacourse.momo.group.repository.CategoryRepository;
-import com.woowacourse.momo.group.service.dto.CategoryResponse;
-
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-@Service
-public class CategoryService {
-
-    private final CategoryRepository categoryRepository;
-
-    public List<CategoryResponse> findAll() {
-        List<Category> categories = categoryRepository.findAll();
-
-        return categories.stream()
-                .map(CategoryResponse::from)
-                .collect(Collectors.toList());
-    }
-}
