+++ backend/src/main/java/com/woowacourse/momo/group/service/dto/request/GroupRequest.java
@@ -10,8 +10,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import com.woowacourse.momo.group.domain.Group;
-import com.woowacourse.momo.group.domain.Schedule;
+import com.woowacourse.momo.group.domain.group.Group;
+import com.woowacourse.momo.group.domain.schedule.Schedule;
 
 @Getter
 @NoArgsConstructor
