+++ backend/src/main/java/com/woowacourse/momo/member/domain/MemberRepository.java
@@ -1,0 +1,6 @@
+package com.woowacourse.momo.member.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+}
