+++ backend/src/test/java/com/woowacourse/momo/category/service/CategoryServiceTest.java
@@ -1,0 +1,29 @@
+package com.woowacourse.momo.category.service;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java.util.List;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.jdbc.Sql;
+
+import com.woowacourse.momo.category.service.dto.response.CategoryResponse;
+
+@SpringBootTest
+@Sql("classpath:init.sql")
+class CategoryServiceTest {
+
+    @Autowired
+    private CategoryService categoryService;
+
+    @DisplayName("카테고리 목록을 조회한다")
+    @Test
+    void findAll() {
+        List<CategoryResponse> actual = categoryService.findAll();
+
+        assertThat(actual).hasSize(10);
+    }
+}
