+++ backend/src/test/java/com/woowacourse/momo/common/acceptance/AcceptanceTest.java
@@ -1,0 +1,19 @@
+package com.woowacourse.momo.common.acceptance;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.test.context.SpringBootTest;
+
+import io.restassured.RestAssured;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class AcceptanceTest {
+
+    @Value("${local.server.port}")
+    int port;
+
+    @BeforeEach
+    public void setUp() {
+        RestAssured.port = port;
+    }
+}
